폴더 구조 변경 및 파일 책임 분리

+++ src/main/java/backend/backend/domain/dailyTask/Repository/DailyTaskRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DailyTaskRepository extends JpaRepository<DailyTask, Integer> {
+public interface DailyTaskRepository extends JpaRepository<DailyTask, Long> {
 }
