+++ src/main/java/backend/backend/domain/joinUser/repository/JoinUserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JoinUserRepository extends JpaRepository<JoinUser, Integer> {
+public interface JoinUserRepository extends JpaRepository<JoinUser, Long> {
     boolean existsByStudentId(String studentId);
 }
